Remake logic
Add a entity "Client"
Remake for english version
Make a few changes in a logic of buying tickets, routes

+++ Cashier_Station/Admin/InterMediateRoute.cs
@@ -17,7 +17,7 @@ namespace Cashier_Station.Admin
         public InterMediateRouteWindow()
         {
             InitializeComponent();
-            FillDropDown();
+            FillDropDownRoute();
         }
 
         private void ExitButton_Click(object sender, EventArgs e)
@@ -25,7 +25,7 @@ namespace Cashier_Station.Admin
             this.Close();
         }
 
-        private void FillDropDown()
+        private void FillDropDownRoute()
         {
             List<string> routeid = new List<string>();
             try
@@ -46,7 +46,7 @@ namespace Cashier_Station.Admin
             }
             catch (Exception ex) 
             {
-                MessageBox.Show("При завантаженні даних з бази даних виникла помилка", "Помилка", MessageBoxButtons.OK, MessageBoxIcon.Error); 
+                MessageBox.Show("An error occurred while loading data from the database", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error); 
                 Console.WriteLine(ex.Message);
             }
             finally 
@@ -64,42 +64,50 @@ namespace Cashier_Station.Admin
 
         private void SubmitButton_Click(object sender, EventArgs e)
         {
-            var result = MessageBox.Show("Ви впевненні, що хочете додати цю інформацію?", "Додавання інформації",
+            var result = MessageBox.Show("Are you sure you want to add this information?", "Adding information",
                 MessageBoxButtons.YesNo, MessageBoxIcon.Question);
 
             if (result == DialogResult.Yes)
             {
                 if (NameTransportTextBox.Text == "" || DistanceFromStart.Text == "" || DistanceToEnd.Text == "")
                 {
-                    MessageBox.Show("Поля не можуть бути пустими або від'ємними", "Помилка", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    MessageBox.Show("Fields cannot be empty or negative", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
                 }
                 else
                 {
                     db.OpenConnection();
                     try
                     {
+                        intermediateroute newIntermediateroute = new intermediateroute
+                        {
+                            RouteId = int.Parse(Routeid.selectedValue),
+                            NameStation = NameTransportTextBox.Text,
+                            DistanceFromStart = double.Parse(DistanceFromStart.Text),
+                            DistanceToEnd = double.Parse(DistanceToEnd.Text),
+                            ArrivalTime = DateTime.Parse(ArrivalTime.Value.ToString())
+                        };
                         string query = "INSERT INTO intermediateroute (RouteId, NameStation, DistanceFromStart, DistanceToEnd, ArrivalTime) VALUES (@RouteId, @NameStation, @DistanceFromStart, @DistanceToEnd, @ArrivalTime)";
 
                         using (MySqlCommand cmd = new MySqlCommand(query, db.GetConnection()))
                         {
-                            cmd.Parameters.AddWithValue("@RouteId", int.Parse(Routeid.selectedValue));
-                            cmd.Parameters.AddWithValue("@NameStation", NameTransportTextBox.Text);
-                            cmd.Parameters.AddWithValue("@DistanceFromStart", int.Parse(DistanceFromStart.Text));
-                            cmd.Parameters.AddWithValue("@DistanceToEnd", int.Parse(DistanceToEnd.Text));
-                            cmd.Parameters.AddWithValue("@ArrivalTime", DateTime.Parse(ArrivalTime.Value.ToString()));
+                            cmd.Parameters.AddWithValue("@RouteId", newIntermediateroute.RouteId);
+                            cmd.Parameters.AddWithValue("@NameStation", newIntermediateroute.NameStation);
+                            cmd.Parameters.AddWithValue("@DistanceFromStart", newIntermediateroute.DistanceFromStart);
+                            cmd.Parameters.AddWithValue("@DistanceToEnd", newIntermediateroute.DistanceToEnd);
+                            cmd.Parameters.AddWithValue("@ArrivalTime", newIntermediateroute.ArrivalTime);
 
                             cmd.ExecuteNonQuery();
                         }
 
-                        Console.WriteLine("Дані успішно додано");
-                        MessageBox.Show("Дані успішно додано", "Успіх", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        Console.WriteLine("Data added successfully");
+                        MessageBox.Show("Data added successfully", "Success", MessageBoxButtons.OK, MessageBoxIcon.Information);
 
                     }
                     catch (Exception ex)
                     {
-                        Console.WriteLine("При додаванні даних у таблицю intermediateroute виникла помилка");
-                        Console.WriteLine($"Помилка: {ex.Message}");
-                        MessageBox.Show("Дані не були додані до бази даних", "Помилка", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                        Console.WriteLine("An error occurred while adding data to the 'intermediateroute' table");
+                        Console.WriteLine($"Error: {ex.Message}");
+                        MessageBox.Show("Data has not been added to the database", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
                     }
                     finally 
                     { 
@@ -111,7 +119,7 @@ namespace Cashier_Station.Admin
 
         private void DeleteButton_Click(object sender, EventArgs e)
         {
-            var result = MessageBox.Show("Ви впевненні, що хочете видалення цю інформацію?", "Видалення інформації",
+            var result = MessageBox.Show("Are you sure you want to delete this information?", "Deleting information",
                 MessageBoxButtons.YesNo, MessageBoxIcon.Question);
             if (result == DialogResult.Yes)
             {
@@ -120,29 +128,84 @@ namespace Cashier_Station.Admin
                 try
                 {
                     db.OpenConnection();
-                    string query = "DELETE FROM intermediateroute WHERE NameStation = @NameStation";
+                    string checkQuery = "SELECT COUNT(*) FROM intermediateroute WHERE NameStation = @NameStation";
 
-                    using (MySqlCommand cmd = new MySqlCommand(query, db.GetConnection()))
+                    using (MySqlCommand checkCmd = new MySqlCommand(checkQuery, db.GetConnection()))
                     {
-                        cmd.Parameters.AddWithValue("@NameStation", NameStationDelete.Text);
+                        checkCmd.Parameters.AddWithValue("@NameStation", StationNameDel.selectedValue);
 
-                        cmd.ExecuteNonQuery();
+                        int rowCount = Convert.ToInt32(checkCmd.ExecuteScalar());
+
+                        if (rowCount > 0)
+                        {
+                            // Запис існує, тепер можна видалити
+                            string deleteQuery = "DELETE FROM intermediateroute WHERE NameStation = @NameStation";
+
+                            using (MySqlCommand deleteCmd = new MySqlCommand(deleteQuery, db.GetConnection()))
+                            {
+                                deleteCmd.Parameters.AddWithValue("@NameStation", StationNameDel.selectedValue);
+
+                                deleteCmd.ExecuteNonQuery();
+                            }
+
+                            Console.WriteLine("The removal was successful");
+                            MessageBox.Show("The removal was successful", "Success", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        }
+                        else
+                        {
+                            // Запис не існує, вивести відповідне повідомлення
+                            MessageBox.Show("Record not found. Nothing was deleted.", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        }
                     }
-
-                    Console.WriteLine("Видалення відбулося успішно");
-                    MessageBox.Show("Видалення відбулося успішно", "Успіх", MessageBoxButtons.OK, MessageBoxIcon.Information);
 
                 }
                 catch (Exception ex)
                 {
-                    Console.WriteLine("При видаленні даних з таблиці intermediateroute виникла помилка");
-                    Console.WriteLine($"Помилка: {ex.Message}");
-                    MessageBox.Show("При видаленні даних з таблиці intermediateroute виникла помилка", "Помилка", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    Console.WriteLine("An error occurred while deleting data from the 'intermediateroute' table");
+                    Console.WriteLine($"Error: {ex.Message}");
+                    MessageBox.Show("An error occurred while deleting data from the 'intermediateroute' table", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
                 }
                 finally
                 {
                     db.CloseConnection();
                 }
+            }
+        }
+
+        private void RouteIdDelete_onItemSelected(object sender, EventArgs e)
+        {
+            try
+            {
+                db.OpenConnection();
+
+                string query = "SELECT NameStation FROM intermediateroute WHERE RouteId = @RouteId";
+
+                using (MySqlCommand cmd = new MySqlCommand(query, db.GetConnection()))
+                {
+                    cmd.Parameters.AddWithValue("@RouteId", RouteIdDelete.selectedValue);
+
+                    using (MySqlDataReader reader = cmd.ExecuteReader())
+                    {
+                        
+                        StationNameDel.Clear();
+
+                        while (reader.Read())
+                        {
+                            string stationName = reader["NameStation"].ToString();
+                            StationNameDel.AddItem(stationName);
+                        }
+                    }
+                }
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine("An error occurred while loading intermediate stations");
+                Console.WriteLine($"Error: {ex.Message}");
+                MessageBox.Show("An error occurred while loading intermediate stations", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+            finally
+            {
+                db.CloseConnection();
             }
         }
     }
